♻️ refactor(quiz): fix UserId import paths for architectural consistency
- Update complete-quiz route and handler to import UserId from user domain
- Aligns with established pattern used in QuizCompletionService
- Improves domain boundary separation and follows DDD principles
- Maintains type compatibility while fixing architectural inconsistency
